--- src/i18n.js
@@ -1,26 +1,0 @@
-import i18n from 'i18next'
-import { initReactI18next } from 'react-i18next'
-import LanguageDetector from 'i18next-browser-languagedetector'
-
-import en from './locales/en/translation.json'
-import ru from './locales/ru/translation.json'
-import hy from './locales/am/translation.json'
-
-const resources = {
-  en: { translation: en },
-  ru: { translation: ru },
-  hy: { translation: hy },
-}
-
-i18n
-  .use(LanguageDetector)
-  .use(initReactI18next)
-  .init({
-    resources,
-    fallbackLng: 'en',
-    interpolation: {
-      escapeValue: false,
-    },
-  })
-
-export default i18n
